refactor: rename models back into types
